put items and collidables into their own files, decrementHealth is now based on impact, and not decrementing an int each time
