+++ GalagaRemake/GameObject.h
@@ -1,7 +1,6 @@
 #pragma once
 #include <SFML/Graphics.hpp>
 #include "PlayerShip.h"
-#include "ItemType.h"
 
 
 class GameObject :
@@ -50,51 +49,3 @@ private:
     void rotateObject();
     void oscillateObject();
 };
-
-class Item :
-	public GameObject
-{
-public:
-    ItemType getItemType() const;
-    void setItemType(ItemType type);
-
-private:
-    ItemType m_itemType = ItemType::Repair_Kit;
-
-};
-
-class Collidable :
-	public GameObject
-{
-public:
-    void setHealth(int health);
-    int getHealth() const;
-    sf::Vector2f getMomentum() const;
-    void setColor(const sf::Color& color);
-    void decrementHealth();
-    void applyMomentum(const sf::Vector2f momentum);
-    void applyMomentum(sf::Vector2f momentum, sf::Vector2f impactLocation);
-    void applyAngularVelocity(float angularVelocity, sf::Vector2f impactLocation, sf::Vector2f momentum);
-	void update() override;
-    std::optional<CollisionResult> detectProjectileCollision();
-    void setMass(float mass);
-    float getAngularVelocity() const;
-    void applyPhysicsToEachOther(Collidable& collidable, const sf::Vector2f pointOfImpact);
-    std::optional<sf::Vector2f> detectCollision(const Collidable& collidable) const;
-    std::optional<sf::Vector2f> detectCollision(const sf::RectangleShape& rect) const;
-    std::optional<sf::Vector2f> detectCollision(const sf::Sprite& sprite) const;
-    void applyPhysicsFromProjectile(CollisionResult collisionResult);
-    int last_collision;
-
-private:
-    void updateObjectHitTimer();
-    void explode();
-    sf::Transform getTransform() const;
-
-
-	int m_health = 1;
-    float m_mass = 1.0;//baseline mass
-    int m_objectHitTimer = -1;
-
-};
-
